requirements.txt
mcp
fastmcp
python-dotenv
requests
gradio
ollama
pika
